Added timed order functionality. Need to write specs.

+++ app/controllers/spree/orders_controller_decorator.rb
@@ -1,12 +1,12 @@
 Spree::OrdersController.class_eval do
-  before_filter :ensure_not_expired, :only => [:edit, :update]
+  before_filter :ensure_not_expired, :only => [:edit, :update, :populate]
   
   private
   
   def ensure_not_expired
-    order = current_order(true)
-    if order.expired? && !order.line_items.empty?
-      flash.notice = "Your order has expired and we released all products in your cart."
+    order = current_order
+    if !order.nil? && order.expired? && !order.line_items.empty?
+      flash.now[:notice] = "Your order has expired and all items in your cart has been released."
       order.line_items.destroy_all
     end
   end
